Update directory structure and package name
