+++ ballerina/Ballerina.toml
@@ -5,12 +5,6 @@ version = "@toml.version@"
 
 [[platform.java11.dependency]]
 path = "../task-native/build/libs/task-native-@project.version@.jar"
-groupId = "ballerina"
-artifactId = "task"
-version = "@project.version@"
 
 [[platform.java11.dependency]]
 path = "./lib/quartz-@quartz.version@.jar"
-groupId = "org.quartz-scheduler"
-artifactId = "quartz"
-version = "@quartz.version@"
