+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.caching=true
-group=org.ballerinalang
+group=io.ballerina.stdlib
 version=2.0.0-beta.3-SNAPSHOT
 ballerinaLangVersion=2.0.0-beta.2.1
 
@@ -10,4 +10,8 @@ quartzVersion=2.3.2
 slf4jVersion=1.7.30
 axiomVersion=1.2.22
 puppycrawlCheckstyleVersion=8.18
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.0
+githubSpotbugsVersion=4.0.5
+githubJohnrengelmanShadowVersion=5.2.0
+underCouchDownloadVersion=4.0.4
+researchgateReleaseVersion=2.8.0
