+++ native/src/main/java/io/ballerina/stdlib/task/actions/TaskActions.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.stdlib.task.actions;
+package io.ballerina.stdlib.task.actions;
 
 import io.ballerina.runtime.api.Environment;
 import io.ballerina.runtime.api.creators.ValueCreator;
@@ -25,10 +25,10 @@ import io.ballerina.runtime.api.values.BDecimal;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
-import org.ballerinalang.stdlib.task.exceptions.SchedulingException;
-import org.ballerinalang.stdlib.task.objects.TaskManager;
-import org.ballerinalang.stdlib.task.utils.TaskConstants;
-import org.ballerinalang.stdlib.task.utils.Utils;
+import io.ballerina.stdlib.task.exceptions.SchedulingException;
+import io.ballerina.stdlib.task.objects.TaskManager;
+import io.ballerina.stdlib.task.utils.TaskConstants;
+import io.ballerina.stdlib.task.utils.Utils;
 import org.quartz.JobDataMap;
 import org.quartz.Scheduler;
 import org.quartz.SchedulerException;
